IIAs test update

+++ ewp-reference-connector/src/main/java/eu/erasmuswithoutpaper/iia/control/IiasEJB.java
@@ -188,7 +188,7 @@ public class IiasEJB {
         foundIia.setHashPartner(iiaInternal.getHashPartner());
         foundIia.setIiaCode(iiaInternal.getIiaCode());
 
-        /*if(foundIia.getCooperationConditions() != null) {
+        if(foundIia.getCooperationConditions() != null) {
             for (CooperationCondition cc : foundIia.getCooperationConditions()) {
                 em.merge(cc);
                 em.remove(cc);
@@ -205,10 +205,10 @@ public class IiasEJB {
                 em.flush();
                 foundIia.getCooperationConditions().add(cc);
             }
-        }*/
+        }
 
 
-        List<CooperationCondition> cooperationConditions = iiaInternal.getCooperationConditions();
+        /*List<CooperationCondition> cooperationConditions = iiaInternal.getCooperationConditions();
         //List<CooperationCondition> cooperationConditionsCurrent = foundIia.getCooperationConditions();//cc in database
         for (CooperationCondition cc : cooperationConditions) {
             String ccId = foundIia.getCooperationConditions().stream().filter(c -> c.getSendingPartner().getInstitutionId().equals(cc.getSendingPartner().getInstitutionId()) && c.getReceivingPartner().getInstitutionId().equals(cc.getReceivingPartner().getInstitutionId())).map(CooperationCondition::getId).findFirst().orElse(null);
@@ -298,7 +298,7 @@ public class IiasEJB {
                     //}
                 //}
             //}
-        }
+        }*/
 
         //foundIia.setCooperationConditions(cooperationConditionsCurrent);
